クラスを統合して、クラスをファイルに切り出す
