+++ lib/procon_bypass_man/device_connection/procon_less.rb
@@ -7,3 +7,4 @@ require_relative "procon_less/output_report_watcher"
 require_relative "procon_less/input_report_builder"
 require_relative "procon_less/output_report_parser"
 require_relative "procon_less/output_report"
+require_relative "procon_less/report_negotiator"
